nat: extract pool-NAT/allocator/sync tests into tests_pool.rs (#4409)
Move the 72 pool source-NAT tests -- pool_snat_* (single/round-robin/
subnet expansion, ICMP query-id distribution, address-persistent sticky
selection, port collision probing/recycle FIFO), the persistent-NAT
lease state machine (reuse/timeout/rollback/reactivation/GC/expiry
invariants), port_allocator_basic, and the four synced-session pool
port-reservation tests (synced_session_*_4388) -- verbatim into
tests_pool.rs, together with every pool test-local helper.

The synced-session tests are grouped here (not in a separate sync file)
because they share the pool port-reservation subject AND the
session_key_from_src() helper with the pool tests; keeping them together
keeps that helper self-contained in one file, so no shared helper has to
be promoted to a common module or duplicated.

Pure code motion: no test body edited, nothing renamed. nat/tests.rs
now holds 46 tests.
