+++ userspace-dp/src/nat/mod.rs
@@ -41,6 +41,9 @@ mod tests_static;
 #[cfg(test)]
 #[path = "tests_destination.rs"]
 mod tests_destination;
+#[cfg(test)]
+#[path = "tests_pool.rs"]
+mod tests_pool;
 
 /// #3096: per-flow interface / routing-instance identity passed into the NAT
 /// match path so an interface- or routing-instance-scoped rule-set matches
